al_ui/requirements.txt
python-baseconv
werkzeug
python-socketio<4
flask
flask-socketio
gevent
gevent-websocket
python_u2flib_server
pyqrcode
markdown
contextvars
elastic-apm[flask]
assemblyline
assemblyline-core